fix error in non-LAPACK version
git-svn-id: https://svn.dealii.org/trunk@13977 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/lac/include/lac/lapack_templates.h
@@ -133,7 +133,7 @@ gemv (const char* trans, const int* m, const int* n, const float* alpha, const f
 }
 #else
 inline void
-gemv (const char*, const int*, const int*, const double*, const double*, const int*, const double*, const int*, const double*, double*, const int*)
+gemv (const char*, const int*, const int*, const float*, const float*, const int*, const float*, const int*, const float*, float*, const int*)
 {
   LAPACKSupport::ExcMissing("sgemv");
 }
@@ -163,7 +163,7 @@ getrf (const int* m, const int* n, float* A, const int* lda, int* ipiv, int* inf
 }
 #else
 inline void
-getrf (const int*, const int*, double*, const int*, int*, int*)
+getrf (const int*, const int*, float*, const int*, int*, int*)
 {
   LAPACKSupport::ExcMissing("sgetrf");
 }
@@ -193,7 +193,7 @@ getrs (const char* trans, const int* n, const int* nrhs, const float* A, const i
 }
 #else
 inline void
-getrs (const char*, const int*, const int*, const double*, const int*, const int*, double*, const int*, int*)
+getrs (const char*, const int*, const int*, const float*, const int*, const int*, float*, const int*, int*)
 {
   LAPACKSupport::ExcMissing("sgetrs");
 }
@@ -223,7 +223,7 @@ geev (const char* jobvl, const char* jobvr, const int* n, float* A, const int* l
 }
 #else
 inline void
-geev (const char*, const char*, const int*, double*, const int*, double*, double*, double*, const int*, double*, const int*, double*, const int*, int*)
+geev (const char*, const char*, const int*, float*, const int*, float*, float*, float*, const int*, float*, const int*, float*, const int*, int*)
 {
   LAPACKSupport::ExcMissing("sgeev");
 }
@@ -253,7 +253,7 @@ geevx (const char* balanc, const char* jobvl, const char* jobvr, const char* sen
 }
 #else
 inline void
-geevx (const char*, const char*, const char*, const char*, const int*, double*, const int*, double*, double*, double*, const int*, double*, const int*, int*, int*, double*, double*, double*, double*, double*, const int*, int*, int*)
+geevx (const char*, const char*, const char*, const char*, const int*, float*, const int*, float*, float*, float*, const int*, float*, const int*, int*, int*, float*, float*, float*, float*, float*, const int*, int*, int*)
 {
   LAPACKSupport::ExcMissing("sgeevx");
 }
@@ -283,7 +283,7 @@ gesvd (int* jobu, int* jobvt, const int* n, const int* m, float* A, const int* l
 }
 #else
 inline void
-gesvd (int*, int*, const int*, const int*, double*, const int*, double*, double*, const int*, double*, const int*, double*, const int*, int*)
+gesvd (int*, int*, const int*, const int*, float*, const int*, float*, float*, const int*, float*, const int*, float*, const int*, int*)
 {
   LAPACKSupport::ExcMissing("sgesvd");
 }
@@ -313,7 +313,7 @@ stev (const char* jobz, const int* n, float* d, float* e, float* z, const int* l
 }
 #else
 inline void
-stev (const char*, const int*, double*, double*, double*, const int*, double*, int*)
+stev (const char*, const int*, float*, float*, float*, const int*, float*, int*)
 {
   LAPACKSupport::ExcMissing("sstev");
 }
